kola: turn --no-default-checks into --allow-default-checks-failures
This switch is still new, so let's just rename it without providing an
alias. The end goal is the same (to allow not failing on broken systemd
units and SELinux violations), but the difference is that we still do
the checks and just log it if failures are found.

This allows humans looking at the logs to verify that only what they
expected to fail actually failed. (For CI/pipelines, long-term I think
we should instead teach kola how to allow only specific failures through
via `kola-denylist.yaml`.)