kola: Do failed units/SELinux checks both before *and* after tests
First: Revert "Revert "kola: Add --no-default-checks""
 (this reverts commit: 57b3520ab1cefaab9b872d0f2f8dce04668dfbff )

Now, when I went to add the SELinux denials check, I completely missed
that we only were checking for denials *before* the test runs.

We obviously want to check both before and after - the same
way we do with e.g. looking for errors on the console.

As part of this, I'd also added `--no-default-checks` but it
actually didn't work.  Fix that so that it does work, and add
a test case with a failing systemd unit that verifies this,
so we can test the test system.

In order to help ratchet things, also add a special `default-checks`
entry to the denylist.  This way if e.g. we are hitting SELinux AVC
denials just on one arch/test we can turn them off temporarily.