tree: rename manifest_lock_overrides to manifest_lock_arch_overrides
Prep for making `manifest_lock_overrides` refer to the archless kind.
(Why not just keep this one named the same and make the new one have a
different name? Because I think the new one will be used the most
commonly by far, so I want it to have the simpler name.)