Add support for arch-independent lockfile overrides
In FCOS, we use "override" lockfiles to pin packages to certain
versions. Right now, we have separate overrides for each base arch we
(eventually want to) support. But that makes maintaining the overrides
cumbersome because of all the duplication.

Now that rpm-ostree supports this, let's make use of it. Add support for
a `manifest-lock.overrides.yaml` arch-independent override file.