Gangplank: make stamp checks better on s390x
This add a second delay in the stamping tests for remote files as well
as more debug information. During CI tests on s390x, the newness check
is failing for some unknown reason; its is hoped that a second dealy and
more debug information will aide in understand the fault.

Signed-off-by: Ben Howard <ben.howard@redhat.com>