Gangplank: fix broken ctrl-c interrupt and OCP3 watch
This fixes two problems:
- When running via podman, a ctrlc-c keyboard interrupt causes Gangplank to hang.
- When running on OpenShift v3.11, the watch.Events Kubernetes API does
  will reported exited pods as running.

ocp.buildconfig.Exec()s go-routine was changed to use a buffered channel for
capturing signals, added SIGHUP to the intercepted signals, and finally,
instead of returning when termination was signaled, it uses a
termination function.

ocp.clusterRunner() was refactored to address differences between OCP3
and OCP4 cluster's with the watch.Events interface. On OCP4, the Pod status
reflects the container states. On OCP3 a pod may show as Pending or Running even
though the individual containers have exited. The switch logic looking
at individual pod status was replaced with individiaul container
evaluation.