mantle: openstack: use clientconfig when creating service clients
The clientconfig library knows how to take into account the settings
in clouds.yaml when creating the service clients. Previously we were
using the openstack NewComputeV2, NewImageServiceV2, NewNetworkV2,
functions directly, which, if not narrowed down appropriately can
match multiple endpoints and will just return the first one in the
list. In the case of the bug, we were ending up talking to different
regions.

Fixes: https://github.com/coreos/coreos-assembler/issues/1977