Gangplank: support uploads for custom build strategy
In the near future, Prow will support running COSA/Gangaplank as a
custom build strategy. In order to provide Prow with the expected
output suitable for E2E testing via Prow, this change completes the
Custom Build Strategy API by emitting a container when the environment
indicates.

To support this, when running in a cluster context, Gangplank:
- Trusts the CA certs provided by the Cluster. All cluster pods
  will run init commands to update the CA trust.
- When a pod creates the base artifacts, and the custom build strategy
  defines an output location, `cosa upload-ostree` is called.
- Added envVar to oscontainer.py to turn off TLS verification
- TLS verification for pushing containers is toggable using the new
  JobSpec `job.disable_tls_verification` boolean.

Signed-off-by: Ben Howard <ben.howard@redhat.com>