digitalocean: set uncompressed-{size,sha256}
cmd-compress doesn't process the DO artifact because we special-case gzip
compression in the mutate callback.  But the artifact is still a regular
compressed disk image, so we should still set the uncompressed
size/sha256.  Have the mutate callback do that as well.

Do not apply a similar change to the GCP tar.gz artifact, since the
uncompressed tarball isn't useful on its own.