cmdlib: use new repo-packages for overrides
Use the new rpm-ostree support for repo-packages to make sure that
rpm-ostree picks the packages from our repo. This is even more explicit
than what we had before with the lockfile which just restricts the
NEVRA. In theory we could drop both the lockfile and `cost` hack now and
put the NEVRAs in the `repo-packages` entry, but meh... they don't
really hurt and we can look at doing that in a follow-up.

More importantly, this will also be required for overrides to work
correctly once rpm-ostree supports modules so that we can easily
override modular and non-modular packages alike.