cmd-kola: temporarily stop testing on NVMe
There's a regression in QEMU v5.2 (which is in f34) where you can't set
`bootindex` on NVMe devices:

https://bugzilla.redhat.com/show_bug.cgi?id=1956429

For now, just skip testing on NVMe when using `cosa kola
--basic-qemu-scenarios`. I tried to just drop the property in the case
where the bootindex shouldn't matter, but couldn't get it to work.

We also include a bomb to make sure that once the bug is fixed, we
re-enable it.