mantle: delete targen code
This isn't used in our current code, and as far as I can tell, was never
really used in the original mantle either.

It was added in https://github.com/coreos/mantle/pull/251 but there
doesn't seem to be any additional context there.

Anyway, let's nuke it since for some reason the test there sometimes
fails on Fedora 34.