cmd-* : multi-arch fixes after f34 move.
Fix cmd-buildextend-live for s390x to use mk-s390image instead of mk-s390-cdboot. Apparently there was a bug in cdboot where it would overwrite the kernel with the initrd if the intrd was too large.
Fix cmd-kola to use qemu-system-{arch} instead of qemu-kvm to check for nvme bootindex as qemu-kvm is not present on non x86 arches