build/gcp: Add uncompressed-checksum
I'm working on https://github.com/cgwalters/coreos-diskimage-rehydrator/
and a goal is to regenerate *exactly* the artifacts.  But right
now because we gzip the GCP tarball at build time, we don't
have `uncompressed-sha256` in the metadata.  Which means the
rehydrator can't verify it.

Add a declarative way to express that an artifact needs to be gzipped.