metal: use bootindex=3 for ISO
This is prep for enabling multipath for the disk on which to install
CoreOS from the live ISO. In multipath mode, boot indices 1 and 2 are
already taken.

For all other purposes, it doesn't make a difference if the CD drive is
index 2 or 3 since it just needs to be lower than the primary disk
(which is 1), so just bump it to 3.

(cherry picked from commit 4da4a30bfee636a85f1262ee7549b3d794d0d8fe)