Ganglank: re-enable CGO
Since the gangway binary executes commands, disabling CGO is breaking
all sorts of commands like tar. In a deep dive, I suspect that this is a
problem with the syscall for ForkExec, although I can't pinpoint it.
Regardless, switching from static for Gangway fixes shell-out commands
while preserving the reason for the switch in the first place.

The tl;dr is that we're in for some fun with the recent changes to
container runtimes with permissions.

Signed-off-by: Ben Howard <ben.howard@redhat.com>