build: Parse image.yaml early on and convert to JSON
Conceptually `image.yaml` is the configuration for coreos-assembler
specifically, distinct from `manifest` and `overlay` etc. which
end up being fed into rpm-ostree.

First, this is part of addressing
https://github.com/coreos/rpm-ostree/pull/2842#discussion_r635542478
so that e.g. we can migrate the `name:` into this file instead
of the rpm-ostree manifest.

But I also want to hook other things off here, such as whether
to use the new ostree-container format instead of our tar-of-archive-repo.