Drop support for no longer used `ostree-commit.tar`
This has been reshuffled a few times, and I plan to do so again
using the new `rpm-ostree ex-container export` bits.  Remove
the now long dead `ostree-commit.tar` bits; the schema has required
the `ostree` entry in `meta.json` for a long time now (since FCOS
from the start, and as of OpenShift 4.3+).