Gangplank: full remote podman support
In order to fully utilize the remote podman containers, this commit adds
support for SSH forwarding of Minio and the ability to set the remote
build architecture.

As part of supporting remote podman, the creation of the ephemeral
"/srv" is far-less hacky -- it now uses the container API for the
creation of a volume that is pruned on exit.

Finally, this removes the requirement to pre-pull the COSA container.
Gangplank running against either local or remote Podman will pull the
image if missing.

Example:
        gangplank pod --podman \
                -A base \
                --remote "ssh://draco@horcrux.dev/run/user/1000/podman/podman.sock"

Co-authored-by: Dusty Mabe <dusty@dustymabe.com>
Signed-off-by: Ben Howard <ben.howard@redhat.com>