platform/qemu: switch boot order for multipath disks
switch boot order for multipath disks when running testiso with `qemu-multipath`. there was a regression in the arm64 rhcos pipeline with testiso failing with:

```
+ kola testiso -S --qemu-native-4k --qemu-multipath --scenarios iso-install --output-dir tmp/kola-metal4k
Testing scenarios: [iso-install]
Error: scenario iso-install: switching boot order failed: Could not set bootindex for blkdev: Running QMP command: Device '' not found; <nil>
2021-06-03T08:32:21Z cli: scenario iso-install: switching boot order failed: Could not set bootindex for blkdev: Running QMP command: Device '' not found; <nil>
```

the disk device for multipath was not recognized when switching bootorder causing this to fail.