Gangplank: use bucket and path prefix for remote objects
Prior to this Gangplank supports S3 in theory -- that is if S3 region
was empty, or in the extremely unlikely event that the caller has the
buckets 'builds' and 'cache.' The reason it worked with Minio is that
the Minio instance was neither shared nor did it remain after the build.

This change allows for Ganplank to use S3 or shared Minio backends by:
- Moving all minio specific information into its own struct
- Requires a bucket, and supports a path prefix for using sub-paths
  inside a bucket.
- Supports S3 secure connections.

It worth highlighing that this change removes the stict requirement of a
builds directory/path for the origin Gangplank. At the point of orgin,
Ganglank will resolve the relative paths and the workers will use the
strict pathing.

Signed-off-by: Ben Howard <ben.howard@redhat.com>