Gangplank: use pure-Go SSH forwarding
Hack-day project: the use of shell-SSH has bothered me -- it was a hack.
For hack-day, I figured de-hacking the SSH to use pure-Go SSH would bea
fitting project.

This borrows from the libpod GoLang implement of SSH code. Since our
needs a different (they use an http client over ssh to a socket) and we
need to proxy, the sshClient function was modified to return an
ssh.Client.

Signed-off-by: Ben Howard <ben.howard@redhat.com>