Mantle/Kola: make qmp.SocketMonitor part of the qemu instance
Previously the qmp socket was hard-coded as a `qmp.sock` in the
instance's tmpdir. This is prone to race conditions (when one VM is torn
down and another is started using the same socket) or worse, cause flaky
and incorrect tests results for multiple VM/cluster scenerios.

This change addresses these problems by:
- adding a `qmp.socketMontior` to the QemuInstance struct
- on machine creation, the socket is opened
- funcs that used the qmp socket are now methods of QemuInstance
- the socket and connection are cleaned-up on instance teardown

Signed-off-by: Ben Howard <ben.howard@redhat.com>