Don't enable `go build -race` by default
This code is new and we're still squashing races. Let's gate it behind
an env var so it's not enabled in production builds since this can cause
failures (see: https://github.com/coreos/coreos-assembler/pull/2219).

This way it's still easy for local devs to enable it and we can still
opt into it for our upstream CI.