create_disk: Add build-kernel
I was reading this code for an unrelated reason and thought
about reproducibility of builds, and I think having the kernel
version here will help.  It's the thing that ends up creating
the actual filesystem layout.

Longer term of course we really need to snapshot version the bits
that go into cosa builds and then reference that snapshot here.