create_disk.sh: remove `options` from s390x parmfile
We should pass a file containing the kargs here, but not the `options`
key we grepped out of the BLS. Otherwise, I think we'll literally end up
with an `options` karg (which thankfully should have been harmless).
Filter through `cut` to remove that.

We regressed on this at some point. The original commit which added
s390x support didn't do this:

https://github.com/coreos/coreos-assembler/commit/100c2e512ecb89786a53bfb1c81abc003776090d