kola/tests/misc/network: Add test to check dhcp prop. over bond via veth
Add a new ci test, that will check dhcp propagation on an ovs-bridge
connected to a bond using veth pairs.

This test creates 2 veth pairs, each with one end set on another network
namespace "test-netns".
The host side of the veth ends get a specific MAC-address
and the netns side veth ends get a specific IP-Address.
The host-side veth ends then receive an IP-Address from a dnsmasq
service running on the netns side.
Finally, we run the setup-ovs script to create a bond and ovs-bridge and
check that the expected MAC and IP are set on the ovs-interface.

Signed-off-by: Ram Lavi <ralavi@redhat.com>