mantle/platform/conf: don't panic on additions to an empty config
If the userdata is specified as Empty(), the test wants to ensure that
there is no Ignition config at all, and all of the additions made to the
config by BaseCluster.RenderUserData() should be ignored.  The panics
added in 22f725a39e18 interfere with that, so remove them again.

Currently, the only test that uses Empty() is fcos.ignition.misc.empty,
which relies on Afterburn SSH key injection.  When that test was ported
from CL in 6e9e49bb12ad, the UserData field wasn't renamed to UserDataV3,
and the nil UserDataV3 caused the test to start running with a default
non-empty Ignition config.  The SSH key was thus injected twice (in the
Ignition config and via Afterburn), but the test doesn't check for that,
so it continued to pass.  Thus, when 22f725a39e18 broke Empty(), the test
continued to work until 1952a49f90df dropped the separate UserDataV3 field
and the test started invoking Empty() again.  Even then, the test passed
cosa CI because it's skipped on qemu, which doesn't have a separate
mechanism for injecting SSH keys.

Fixes panic in fcos.ignition.misc.empty:

    panic: Could not find a supported Ignition config version
    goroutine 5056 [running]:
    github.com/coreos/mantle/platform/conf.(*Conf).AddAuthorizedKeys(0xc000125200, 0x20e06cb, 0x4, 0xc00086ab50, 0x1, 0x1)
    	/root/containerbuild/mantle/platform/conf/conf.go:690 +0x16b
    github.com/coreos/mantle/platform/conf.(*Conf).CopyKeys(0xc000125200, 0xc000010a10, 0x1, 0x1)
    	/root/containerbuild/mantle/platform/conf/conf.go:701 +0x165
    github.com/coreos/mantle/platform.(*BaseCluster).RenderUserData(0xc000735900, 0xc000110f30, 0xc000923d58, 0xd, 0xc001435e20, 0x40)
    	/root/containerbuild/mantle/platform/cluster.go:207 +0x6c5
    github.com/coreos/mantle/platform/machine/aws.(*cluster).NewMachine(0xc00086ab10, 0xc000110f30, 0x0, 0x0, 0x0, 0xc001435ef8)
    	/root/containerbuild/mantle/platform/machine/aws/cluster.go:32 +0x16d
    github.com/coreos/mantle/platform/machine/aws.(*cluster).NewMachineWithOptions(0xc00086ab10, 0xc000110f30, 0x0, 0x0, 0x0, 0x0, 0x1debd01, 0xc001435fa8, 0x30000000056545a, 0xc000ce9740)
    	/root/containerbuild/mantle/platform/machine/aws/cluster.go:85 +0x93
    github.com/coreos/mantle/platform.NewMachines.func1(0xc0015ad1b0, 0x244ea68, 0xc00086ab10, 0xc000110f30, 0x0, 0x0, 0x0, 0x0, 0xc001b40d80, 0xc001b40d20)
    	/root/containerbuild/mantle/platform/platform.go:364 +0xcb
    created by github.com/coreos/mantle/platform.NewMachines
    	/root/containerbuild/mantle/platform/platform.go:362 +0x165