Mantle: use events for devshell
This refactor makes devshell a bit smarter by using systemd events and
the serial log to determine events. When using the devshell during
reboot, halt, and shutdown the developer might not know what's
happening.

The systemd eventing detection is really best-effort. Events from
systemd might not be flushed when/if the virtio job is done.