buildextend-*: Add `--compress`
The RHCOS pipeline is trying to add an `azurestack` artifact and
that's blowing transient disk usage past the the 50Gi PVC.

There's no real reason not to compress right after building
artifacts that we're not going to immediately test locally (e.g.
like we do with `qemu`) so add an option to compress directly.