kola: Add RunCmdSync that logs cmd to target journal, use in upgrade test
I'm trying to debug a failure in the upgrade test and not getting
any logging around the invocations of commands makes it really hard.

Because we always capture the target system journal by default,
add a new wrapper for `MustSSH` that prints the command to the
journal too.

I considered changing `MustSSH` by default but that's used
in a *lot* of places and I'd like to try this out in
just the upgrade test to start.  If this works OK we can
consider a tree-wide change.

(cherry picked from commit ad266b3e88fdad38a173e6a201f57cac531a1e43)