gangplank: add support for uploading extra files
For bump lockfile we want to be able to just harvest off the updated
lockfile (no actual builds happen). In that case let's create the file
under tmp/ with a .minioupload suffix and match on that when we upload
things.

This also changes the code to opt out of uploading build contents if
no builds exist.