Dockerfile: Stop running unit tests
We support building this image in many places; in OpenShift
clusters, locally via `podman` as well as in the special `quay.io`
build environment.  The gangplank tests are failing in the latter.

In practice, don't need to re-run unit tests in quay.io.  Code
landing via PRs is already tested.