kola: add new tracker_list field for denylist
In openshift/os#595 and openshift/os#599, I found myself mixed up with
the support of the `tracker` field.  It's defined as a string, so
magically hoping that a slice of strings would be supported was
misguided.

Decided to try adding support for a new `tracker_list` field for those
times when one issue just doesn't cut it.

This seemed more straight-forward than trying to manually unmarshal
the YAML or trying to find all the denylist documents in use.