kola/testiso.go: ignore signature verification on dev builds
It can be misleading when running a local test that it fails even when technically
nothing is wrong (development builds aren't expected to be signed).
Changed the code so that we ignore signature verification for development builds.
Also, logged for both scenarios where we ignore signature

xref: https://github.com/coreos/fedora-coreos-tracker/issues/908