ci/prow-rhcos.sh: Also map `master` → `master`
In rehearsal I'm seeing
`Unhandled base ref: master` which I *think* may be
due to the fact rehearsal is a very special case that is
simulating a PR, and it doesn't know the default branch is `main`.

Let's just pass through `master` as the identity even though
we shouldn't need it later.