cmd-buildprep: allow for --arch=all
Add support for the user specifying --arch=all, which will inspect
the builds.json for a particular build and download the meta.json
and ostree commit object for all architectures related to that build.

This is useful when doing things like generating the release.json
which needs information from each of the architecture builds in
order to give a complete release.json.