cosalib: support parent_build argument to init_build_meta_json
If there is ever a case where there are missing architectures for
a given build then the pipeline will buildprep the meta.json from
the last released build for that architecture. Now we need to be
able to tell init_build_meta_json that the parent for the given
architecture might not be the latest build, in which case we'll
specify the parent build ID directly.