mantle/kola: Add ability to run non-exclusive tests in one VM
Tests that do not conflict with other tests (non-exclusive) can now be run in
one VM. The main change here is that instead of creating a cluster and n
machines for each test, we launch one cluster and one machine which runs all
the non-exclusive tests as subtests. Reboots are turned off for non-exclusive
tests.