Gangplank: use log API for logs
Prior to this, Gangplank used `containers.Attach` for capturing the logs
from worker containers; this was done since the old Podman API was
nearly unusable for watching logs.

However, using `Attach` will fail while losing logs if the container
exits before the attach completes. Since 3.2 Podman API does not have
the performance penalty, the prudent course is to use the Log API call
to ensure that all logs are captured.

Signed-off-by: Ben Howard <ben.howard@redhat.com>