mantle/Makefile: don't trigger a build on install
Otherwise, if one does e.g. `sudo make install`, we'll dumbly try to
rebuild binaries as root because `build` is a phony target. Ideally it
wouldn't be, but golang has its own change detection logic and it'd be
inefficient to try to replicate it in the Makefile.

Instead, let's just leave it to the user to make sure to build first and
otherwise we simply let the install fail if a binary is missing.