qemuexec: support multipath on additional disks
Right now, `--qemu-multipath` turns on multipathing for the primary disk
and all additional disks. But ideally, we want these to be independent,
so that e.g.  we could turn on multipath just for an additional disk but
not the primary disk (or vice versa).

Change `--qemu-multipath` to make it only apply to the primary disk.
Then, add support for specifying comma-separated options to additional
disks using the format `--add-disk <size>:<opt1>,<opt2>,...`. Add
support for `mpath` as a disk option to mean turning on multipath just
for that disk.

This allows me to easily interactively test how the OS behaves before
writing a test which does the same thing.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=1981973