gangplank: take two on ssh port forwarding
This time we'll go ahead and use the remote chosen port locally as
well. This is because the minio bucket is consulted by the local running
process for required artifacts before even starting a pod on the remote
(i.e. why start the pod until the artifacts exist that it needs?). If
they use different port numbers then we'd have to track them separately
and behave differently whether we're running in the remote podman cosa or
in the local cosa. Let's just use the same port locally as well.

To achieve this goal we submerge the starting of the minio server into
the ssh port forwarding code so we can detect the port before we start
the server and set up the proxy.