gangplank: fix subtle creation of new remoteSSHport var
We need to re-use the one that was created outside of the for loop
not create a newly scoped version of remoteSSHport. This meant the
port got changed to 0:

```
INFO[0000] Changing minio port for local and remote (forward) from 0 to 0
```

and since there was already special handling to assign the port to 9000
if it was 0 then it made it through. Fix this now.