metal: append `rw` karg when testing multipath
It's usually injected by the rootmap on first boot, but if we're booting
with `root=/dev/disk/by-label/dm-mpath-root` from the start, then the
rootmap code doesn't run.

Add it in the testiso path to match the same procedure a user would
follow. This has been passing in CI because we don't reboot the
installed system to verify it comes up fine again (we probably should).