tests/multipath: add multipath.partition test
There have been requests to support multipath on specific partitions
instead of the whole disk. There are some caveats, but this should work.
Add a test for this which verifies that `/var/lib/containers` can be on
multipath.

The huge annoyance here is that we can't use Ignition to format the
disk. I'm hoping we can clean this up eventually, but it's intimately
tied to how multipath itself works and how FCOS/RHCOS integrates it.

Related: https://bugzilla.redhat.com/show_bug.cgi?id=1981973