gangplank: fix out of order execution of commands
The list of commands can be executed in order serially or concurrently.
The in order serial piece was broken because the map used to hold info
about the scripts to execute is unordered. Let's fix it so that the
scripts are executed in the same order they were added to the map, thus
making them execute serially in order.