qemu: use `-nodefaults`
By default, QEMU allocates a bunch of things to all the VMs it
provisions. Since we want to have full control over the VM, let's use
`-nodefaults` to disable this.

If this breaks anything, then we should explicitly add the virtual
hardware that we need.

With this for example, we no longer get a CD drive attached. E.g. right
now a `cosa run` will show you:

```
$ lsblk /dev/sr0
NAME MAJ:MIN RM  SIZE RO TYPE MOUNTPOINT
sr0   11:0    1 1024M  0 rom
```

One concrete improvement from this is that it squashes this GRUB error
we get on boot:

    error: ../../grub-core/disk/i386/pc/biosdisk.c:543:failure
    reading sector 0x1670 from `fd0'.

(The GRUB manual implies that `fd*` is for floppy disk drives, but I'm
not seeing any floppy drives connected to the system when not using
`-nodefaults`, so I think there's some mislabeling going on somehow and
it's actually trying to read from the empty CD drive.)