mantle/platform: handle the x86_64 uefi case for PXE
Needed to add some conditional logic that would set up PXE correctly
for UEFI. Also chose to use `bootindex=2` for the NIC. The first boot
will fallthrough from the hard drive (`bootindex=1`) to the second entry
(NIC/PXE). The second boot will have an OS because the install was done.