cmd-init: support --commit
Useful for reseting the local branch of the recently cloned config
git repo to a specific commit (full SHA-1).

This is most useful in a automated environment where we're doing a
`cosa init` and we might not always want the latest config repo, but
rather a specific commit (which could be latest, or not). One example
of this being useful is when doing pipeline builds and targetting a
specific commit from subordinate multi-arch build pipelines that run
later in time than the x86_64 pipeline. For our devel branches new
commits could have easily been commited in the amount of time between
the x86_64 pipeline ran and the multi-arch pipelines run.