gangplank: allow for specifying configs git commit
This exposes the `cosa init --commit` flag into the jobspec, which
allows one to force a commit on the specified branch to be used. This
is useful in pipeline builds when later subordinate (multi-arch)
pipelines run and we want them to target the same git commit as the
earlier x86_64 pipeline run.