kola: Change RunCmdSync to log command output to journal too
The original goal of this new API was just to log the command invocation
to the journal, but looking at things further here, the common path
is to *not* capture the output.

Fix things here so that we log both the command stdout+stderr
as a unified stream into the journal, and then capture the last 20
lines from the journal for `kola` as a Go error message.

(In practice, we should probably not capture the journal text into
 `t.Fatalf()` but just say "see the journal" and then have a higher
 level pass which gathers relevant failures from the journal
 from all machines)

My immediate motivation is improved debugging for hotfix
failures: https://github.com/coreos/fedora-coreos-tracker/issues/942