kola: Replace all non-capturing MustSSH with RunCmdSync
Probably 70% of the invocations of `MustSSH` weren't actually
capturing stdout, so it would just get lost.  Instead, use
`RunCmdSync` for this which keeps everything consistently in
the systemd journal in order.

This will help debugging because often one needs to consult
the journal, and having log messages there will help.