build: Keep Dockerfile + ci/Dockerfile in sync, require rpm-ostree v2021.10
The internal RHCOS pipeline is rebuilding cosa due to multiarch, and
there was a subtle difference that crept in between the two `Dockerfile`s;
notably the main one wasn't configuring the `-continuous` repo
where we'd bypassed the manual Bodhi stuff to fast track
a new `rpm-ostree`.

We need that for https://github.com/openshift/os/pull/603

Sync the two `Dockerfile`s and also require the updated version.