kola: Add support for .ociarchive to RHCOS upgrade test
I forgot that we forked the upgrade tests between FCOS/RHCOS
when doing https://github.com/coreos/coreos-assembler/pull/2300

The ugly hack here is to downconvert the oci image on the kola
container side to the old tar format, because the rpm-ostree
in RHEL8 is too old to natively accept the ociarchive.

Once that is fixed, (which will be helped by us testing
this end-to-end) then this code will get nicer.