build: Add a `:latest` tag to ociarchive export
For rpm-ostree, we currently need the two colons `:` in e.g.
`rpm-ostree rebase oci-archive:/path/to/blah:latest` to know
that we're rebasing to a container image (as opposed to just
`rpm-ostree rebase oci-archive:/path/to/blah` which may be
an ostree remote named `oci-archive`.

However, unlike the docker registry transport, by default
no tag `:latest` is emitted into the archive unless specified.