mantle: Add `COSA_IGNITION_DEFAULT_VERSION` envvar mainly for `cosa run`
Every time I go to `cosa run rhcos-4.6.qcow2` I keep being confused
by the failure in the initramfs, which is just because we inject
a too-new Ignition version.

I am sure we will want this for FCOS too - being able to conveniently
boot older disk images is very useful for bisection for example.

Ideally we make this configurable/autodetected, but for now
this hidden environment variable is enough to unblock me.