kola/harness.go: fix panic on non-exclusive test fail
fixes: https://github.com/coreos/coreos-assembler/issues/2407

Failure of a non-exclusive test would result in a golang panic. We were
using a TestCluster object that references the parent's harness.H. Thus, on
error the parent's goroutine would incorrectly be terminated.

To fix the problem, we create a new TestCluster for each test. This will
reference the correct harness.H object for its corresponding test.