create_disk: add reserved partition entries to aarch64,ppc64le
We decided to try to keep the partition numbers in sync across our
x86_64, aarch64, and ppc64le platforms. Previously we skipped partition
numbers on some architectures (for example there was no sda1 by default
on aarch64) and this caused a few issues:

- user confusion
    - it is a common misconception that partition numbers need to start
      at one and increment
    - it can be surprising if the disk layout changes when you go from
      one architecture from the other and re-use the same ignition config
- tooling bugs
    - the butane templates for device boot mirroring didn't include
      partition numbers and ended up with inconsistent behavior on
      aarch64/ppc64le versus x86_64

Keeping the partition numbers in sync will help alleviate some of these
issues.

Fixes: https://github.com/coreos/fedora-coreos-tracker/issues/855