mantle/kola/qemu: support attaching ISO as regular disk
We support writing the live ISO to a USB stick and booting that way, but
that path exercises different bootloader code and we should test it
explicitly.  Do so when --qemu-iso-as-disk is specified.  QEMU's UEFI
firmware still tries to boot via El Torito in this scenario, but some
hardware prefers or requires booting via the ESP in the hybrid GPT, so
overwrite the El Torito signature in this case to force QEMU to boot via the
ESP.