kola: Add an AssertCmdOutputContains
Followup to previous work on `RunCmdSync` which is aiming
to clean up how we run processes.

The rationale for this helper is much the same as APIs like
https://pkg.go.dev/github.com/stretchr/testify@v1.7.0/assert#Contains

It encapsulates this common pattern in a single line.

Also unlike `MustSSH`, we log the command to the systemd journal
too.

(cherry picked from commit f361834edf2c1b1b6f5bcacc01666ae925008bc4)