Remove shim fallback.efi; chain directly to grub
CoreOS makes two assumptions unusual for a Linux distro: that it is the
only OS on the machine, and that it is installed by dd'ing a disk image
onto a disk that's booted on interchangeable hardware.  As such, we've
generally assumed that we do not configure UEFI boot variables, and
instead rely on UEFI's fallback boot mechanism to boot the OS.  But in
fact, shim chains to fallback.efi when booted via the fallback mechanism,
and the latter sets up UEFI boot variables for us.  If the machine is
subsequently reprovisioned from a newer metal image (with a different
ESP GUID), this will happen again, leaving a stale boot variable behind.
Some firmware gets confused by this and refuses to boot, and in any case,
doing this enough times could eventually fill up EFI variable storage.

The ISO image also creates boot variables via the same mechanism, which is
even worse because the ISO media will probably be removed from the machine
after one boot.

Delete fallback.efi.  In its place, copy grub.efi into the EFI/BOOT
directory, where the fallback shim can find it and chain to it.  Leave the
existing copy of grub in place, for existing systems booting via boot
variables.