mantle/kola/testiso: split iso-as-disk to separate function
We'll eventually want to add test services, and we can't do that while
sharing a code path with iso-live-login because the latter requires that
no Ignition config is provided to the live system.  Add a new function
for iso-as-disk, and record the journal in that path.

Don't bother creating a target disk in the iso-as-disk scenario, since
we're not installing anything.