buildextend-live: drop shim fallback.efi from ISO; simplify EFI image
UEFI boots from removable media via the arch-specific default EFI
application in /EFI/BOOT.  When booted that way, shim chains to
fallback.efi if it exists, and fallback.efi creates an EFI boot entry.
That's not appropriate for removable media boot, since the media will
probably never be present again.  If a TPM is present, fallback.efi will
additionally reboot the machine, and on some machines this leads to boot
loops.  Instead of all this, we just want shim to chain directly to GRUB.

The EFI boot image currently looks like this:

    EFI/BOOT/BOOTX64.EFI
    EFI/BOOT/fbx64.efi
    EFI/fedora/BOOTX64.CSV
    EFI/fedora/fonts/
    EFI/fedora/grub.cfg
    EFI/fedora/grubx64.efi
    EFI/fedora/mmx64.efi
    EFI/fedora/shim.efi
    EFI/fedora/shimx64.efi

Refactor the EFI boot image under the expectation that we never want to
create or boot from an EFI boot entry.  Consolidate /EFI/<vendor-id> into
/EFI/BOOT and delete things we don't need: fallback.efi, its associated
CSV, and extra copies of shim.  The result is this:

    EFI/BOOT/BOOTX64.EFI
    EFI/BOOT/fonts/
    EFI/BOOT/grub.cfg
    EFI/BOOT/grubx64.efi
    EFI/BOOT/mmx64.efi

This also saves a couple MB of space.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2004449